How to handle props:
- what is feature?
-build the tree folder in a project
- use prop in project
