+++ src/features/MagicBox/components/Box/Box.scss
@@ -1,0 +1,9 @@
+.box{
+    display:inline-block;
+    padding: 20px;
+    border-radius: 5px;
+
+    color: white;
+    font-size: 24px;
+    font-weight: bold;
+}
