+++ src/features/Rendering/pages/index.scss
@@ -1,0 +1,15 @@
+.app{
+    padding: 1rem 2 rem;
+}
+
+.student-list{
+
+}
+
+.color-list{
+    margin-top: 3rem;
+
+    &>li{
+        text-transform: uppercase;
+    }
+}
